pyproject.toml
[tool.poetry]
name = "claude-container"
version = "0.1.0"
description = ""
authors = ["Adam Holsinger <adamholsinger@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
click = "^8.2.1"
docker = "^7.1.0"
pyyaml = "^6.0.2"
pydantic = "^2.0.0"
tabulate = "^0.9.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-mock = "^3.12.0"
pytest-cov = "^5.0.0"

[tool.poetry.scripts]
claude-container = "claude_container.cli.main:cli"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
